Tested code.
