+++ src/ApiService/ApiService/onefuzzlib/WebhookOperations.cs
@@ -55,7 +55,8 @@ public class WebhookOperations : Orm<Webhook>, IWebhookOperations
              Event: eventMessage.Event,
              InstanceId: eventMessage.InstanceId,
              InstanceName: eventMessage.InstanceName,
-             WebhookId: webhook.WebhookId
+             WebhookId: webhook.WebhookId,
+             TryCount: 0
             );
 
         var r = await _webhookMessageLogOperations.Replace(message);
@@ -76,7 +77,7 @@ public class WebhookOperations : Orm<Webhook>, IWebhookOperations
 
         var (data, digest) = await BuildMessage(webhookId: webhook.WebhookId, eventId: messageLog.EventId, eventType: messageLog.EventType, webhookEvent: messageLog.Event, secretToken: webhook.SecretToken, messageFormat: webhook.MessageFormat);
 
-        var headers = new Dictionary<string, string> { { "Content-type", "application/json" }, { "User-Agent", USER_AGENT } };
+        var headers = new Dictionary<string, string> { { "User-Agent", USER_AGENT }};
 
         if (digest != null)
         {
@@ -84,6 +85,7 @@ public class WebhookOperations : Orm<Webhook>, IWebhookOperations
         }
 
         var client = new Request(_httpFactory.CreateClient());
+        _log.Info(data);
         var response = client.Post(url: webhook.Url, json: data, headers: headers);
         var result = response.Result;
         if (result.StatusCode == HttpStatusCode.Accepted)
@@ -107,6 +109,7 @@ public class WebhookOperations : Orm<Webhook>, IWebhookOperations
         {
             var instanceId = await _containers.GetInstanceId();
             var webhookMessage = new WebhookMessage(WebhookId: webhookId, EventId: eventId, EventType: eventType, Event: webhookEvent, InstanceId: instanceId, InstanceName: _creds.GetInstanceName());
+            
             data = JsonSerializer.Serialize(webhookMessage, options: EntityConverter.GetJsonSerializerOptions());
         }
 
@@ -117,7 +120,6 @@ public class WebhookOperations : Orm<Webhook>, IWebhookOperations
             hmac.Key = System.Text.Encoding.UTF8.GetBytes(secretToken);
             digest = Convert.ToHexString(hmac.ComputeHash(System.Text.Encoding.UTF8.GetBytes(data)));
         }
-
         return new Tuple<string, string?>(data, digest);
 
     }
@@ -136,3 +138,165 @@ public class WebhookOperations : Orm<Webhook>, IWebhookOperations
     }
 
 }
+
+public interface IWebhookMessageLogOperations : IOrm<WebhookMessageLog>
+{
+    IAsyncEnumerable<WebhookMessageLog> SearchExpired();
+    public Async.Task ProcessFromQueue(WebhookMessageQueueObj obj);
+}
+
+
+public class WebhookMessageLogOperations : Orm<WebhookMessageLog>, IWebhookMessageLogOperations
+{
+    const int EXPIRE_DAYS = 7;
+    const int MAX_TRIES = 5;
+
+    private readonly IQueue _queue;
+    private readonly ILogTracer _log;
+    private readonly IWebhookOperations _webhook;
+
+    public WebhookMessageLogOperations(IStorage storage, IQueue queue, ILogTracer log, IServiceConfig config, ICreds creds, IHttpClientFactory httpFactory, IContainers containers) : base(storage, log, config)
+    {
+        _queue = queue;
+        _log = log;
+        _webhook = new WebhookOperations(httpFactory: httpFactory, creds: creds, storage: storage, webhookMessageLogOperations: this, containers: containers, log: log, config: config);
+    }
+
+
+    public async Async.Task QueueWebhook(WebhookMessageLog webhookLog)
+    {
+        var obj = new WebhookMessageQueueObj(webhookLog.WebhookId, webhookLog.EventId);
+
+        TimeSpan? visibilityTimeout = webhookLog.State switch
+        {
+            WebhookMessageState.Queued => TimeSpan.Zero,
+            WebhookMessageState.Retrying => TimeSpan.FromSeconds(30),
+            _ => null
+        };
+
+        if (visibilityTimeout == null)
+        {
+            _log.WithTags(
+                    new[] {
+                        ("WebhookId", webhookLog.WebhookId.ToString()),
+                        ("EventId", webhookLog.EventId.ToString()) }
+                    ).
+                Error($"invalid WebhookMessage queue state, not queuing. {webhookLog.WebhookId}:{webhookLog.EventId} - {webhookLog.State}");
+        }
+        else
+        {
+            await _queue.QueueObject("webhooks", obj, StorageType.Config, visibilityTimeout: visibilityTimeout);
+        }
+    }
+
+    public async Async.Task ProcessFromQueue(WebhookMessageQueueObj obj)
+    {
+        var message = await GetWebhookMessageById(obj.WebhookId, obj.EventId);
+
+        if (message == null)
+        {
+            _log.WithTags(
+                new[] {
+                    ("WebhookId", obj.WebhookId.ToString()),
+                    ("EventId", obj.EventId.ToString()) }
+            ).
+            Error($"webhook message log not found for webhookId: {obj.WebhookId} and eventId: {obj.EventId}");
+        }
+        else
+        {
+            await Process(message);
+        }
+    }
+
+    private async System.Threading.Tasks.Task Process(WebhookMessageLog message)
+    {
+
+        if (message.State == WebhookMessageState.Failed || message.State == WebhookMessageState.Succeeded)
+        {
+            _log.WithTags(
+                new[] {
+                    ("WebhookId", message.WebhookId.ToString()),
+                    ("EventId", message.EventId.ToString()) }
+            ).
+            Error($"webhook message already handled. {message.WebhookId}:{message.EventId}");
+            return;
+        }
+
+        var newMessage = message with { TryCount = message.TryCount + 1 };
+
+        _log.Info($"sending webhook: {message.WebhookId}:{message.EventId}");
+        var success = await Send(newMessage);
+        if (success)
+        {
+            newMessage = newMessage with { State = WebhookMessageState.Succeeded };
+            await Replace(newMessage);
+            _log.Info($"sent webhook event {newMessage.WebhookId}:{newMessage.EventId}");
+        }
+        else if (newMessage.TryCount < MAX_TRIES)
+        {
+            newMessage = newMessage with { State = WebhookMessageState.Retrying };
+            await Replace(newMessage);
+            await QueueWebhook(newMessage);
+            _log.Warning($"sending webhook event failed, re-queued {newMessage.WebhookId}:{newMessage.EventId}");
+        }
+        else
+        {
+            newMessage = newMessage with { State = WebhookMessageState.Failed };
+            await Replace(newMessage);
+            _log.Info($"sending webhook: {newMessage.WebhookId} event: {newMessage.EventId} failed {newMessage.TryCount} times.");
+        }
+
+    }
+
+    private async Async.Task<bool> Send(WebhookMessageLog message)
+    {
+        var webhook = await _webhook.GetByWebhookId(message.WebhookId);
+        if (webhook == null)
+        {
+            _log.WithTags(
+                new[] {
+                    ("WebhookId", message.WebhookId.ToString()),
+                }
+            ).
+            Error($"webhook not found for webhookId: {message.WebhookId}");
+            return false;
+        }
+
+        try
+        {
+            return await _webhook.Send(message);
+        }
+        catch (Exception exc)
+        {
+            _log.WithTags(
+                new[] {
+                    ("WebhookId", message.WebhookId.ToString())
+                }
+            ).
+            Exception(exc);
+            return false;
+        }
+
+    }
+
+    private void QueueObject(string v, WebhookMessageQueueObj obj, StorageType config, int? visibility_timeout)
+    {
+        throw new NotImplementedException();
+    }
+
+    public IAsyncEnumerable<WebhookMessageLog> SearchExpired()
+    {
+        var expireTime = (DateTimeOffset.UtcNow - TimeSpan.FromDays(EXPIRE_DAYS)).ToString("o");
+
+        var timeFilter = $"Timestamp lt datetime'{expireTime}'";
+        return QueryAsync(filter: timeFilter);
+    }
+
+    public async Async.Task<WebhookMessageLog?> GetWebhookMessageById(Guid webhookId, Guid eventId)
+    {
+        var data = QueryAsync(filter: $"PartitionKey eq '{webhookId}' and RowKey eq '{eventId}'");
+
+        return await data.FirstOrDefaultAsync();
+    }
+}
+
